moved async call to pagelayout
