+++ components/Navbar.js
@@ -1,57 +1,12 @@
-import React, { useState, useEffect } from "react";
-import { useSelector, useDispatch } from "react-redux";
-import { clearSignIn, updateSignInError } from "../actions/signIn";
-import apiFetch from "../functions/apiFetch";
+import React from "react";
 import { Colours } from "../definitions";
-import { useRouter } from "next/router";
 import styled from "styled-components";
 import Link from "next/link";
 import Button from "./Button";
 
-const Navbar = ({ className }) => {
-  const signInState = useSelector((state) => state.signIn);
-  const [signedIn, setSignedIn] = useState(false);
-  const dispatch = useDispatch();
-  const router = useRouter();
-
-  const isSignedIn = async () => {
-    try {
-      let response = await apiFetch("/user/session", {});
-      return response.status === 200 ? setSignedIn(true) : setSignedIn(false);
-    } catch (error) {
-      console.log(error);
-      return false;
-    }
-  };
-
-  //Establishes sign-in status to determine visibility of sign-out button
-  useEffect(() => {
-    isSignedIn();
-  }, [signInState]);
-
-  const handleSignout = async () => {
-    let response = await apiFetch("/user/session", {
-      body: signInState.body,
-      method: "DELETE",
-    });
-
-    if (response.status === 200) {
-      dispatch(clearSignIn());
-      setSignedIn(false);
-      router.push("/signin");
-    }
-
-    //TODO: maybe else here for something
-    /**
-     *           <Alert
-            message={signInState.alerts.error}
-            onClose={() => dispatch(clearSignInAlerts())}
-          />
-     */
-  };
-
+const Navbar = ({ ...props }) => {
   return (
-    <Header className={className}>
+    <Header className={props.className}>
       <HeaderContents>
         <Link href="/">
           <img className="headerLogo" src="/img/todox-logo-white.svg" />
@@ -61,9 +16,7 @@ const Navbar = ({ className }) => {
           text="Sign out"
           size="large"
           variant="primary"
-          onClick={handleSignout}
-          disabled={!signedIn}
-          hideAll={!signedIn}
+          {...props}
         />
       </HeaderContents>
     </Header>
