+++ components/PageLayout.js
@@ -1,10 +1,46 @@
-import React from "react";
-import Head from "next/head";
-import styled from "styled-components";
+import React, { useState, useEffect } from "react";
+import { useSelector, useDispatch } from "react-redux";
+import { clearSignIn } from "../actions/signIn";
+import apiFetch from "../functions/apiFetch";
 import { Colours } from "../definitions";
+import { useRouter } from "next/router";
+import styled from "styled-components";
+import Head from "next/head";
 import Navbar from "./Navbar";
 
 const PageLayout = ({ className, title, children }) => {
+  const signInState = useSelector((state) => state.signIn);
+  const [signedIn, setSignedIn] = useState(false);
+  const dispatch = useDispatch();
+  const router = useRouter();
+
+  const isSignedIn = async () => {
+    try {
+      let response = await apiFetch("/user/session", {});
+      return response.status === 200 ? setSignedIn(true) : setSignedIn(false);
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
+  //Establishes sign-in status to determine visibility of sign-out button
+  useEffect(() => {
+    isSignedIn();
+  }, [signInState]);
+
+  const handleSignout = async () => {
+    let response = await apiFetch("/user/session", {
+      body: signInState.body,
+      method: "DELETE",
+    });
+
+    if (response.status === 200) {
+      dispatch(clearSignIn());
+      setSignedIn(false);
+      router.push("/signin");
+    }
+  };
+
   return (
     <Container className={className}>
       <Head>
@@ -12,7 +48,11 @@ const PageLayout = ({ className, title, children }) => {
         <meta httpEquiv="Content-Type" content="text/html; charset=utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
       </Head>
-      <Navbar />
+      <Navbar
+        onClick={handleSignout}
+        disabled={!signedIn}
+        hideAll={!signedIn}
+      />
       <div className="contentContainer">{children}</div>
     </Container>
   );
